% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_af_fluor_fast.R
\name{assign.af.fluor.fast}
\alias{assign.af.fluor.fast}
\title{Assign AF Spectrum By Fluorophore Projection Fast}
\usage{
assign.af.fluor.fast(raw.data, spectra, af.spectra, threads = 1)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{threads}{Numeric, default is \code{1}.}
}
\value{
The indices of the best-fitting AF spectrum per cell
}
\description{
Projects the autofluorescence spectral variants into fluorophore unmixed
space to determine which best fits each cell (event). A fast approximation for
brute force sequential unmixing method in early versions of AutoSpectral.
Provides essentially identical results to minimization of fluorophore signal
(dist0 method). Substantially faster. Use L1 (absolute value) minimization,
which works better than the standard L2 (squared error). Wrapper for the C++
version, which uses OpenMP for parallel processing.
}
