% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_kde2d_cpp}
\alias{fast_kde2d_cpp}
\title{Fast Kernel Density Estimation in 2D}
\usage{
fast_kde2d_cpp(x, y, n, h, x_limits, y_limits)
}
\arguments{
\item{x}{Numeric vector of x-values (typically FSC-A)}

\item{y}{Numeric vector of y-values (typically SSC-A)}

\item{n}{Number of grid points for both x and y (recommend 100)}

\item{h}{Vector of bandwidths for x and y (recommed 0.1*bandwidth.nrd)}

\item{x_limits}{Limits of the data in the x coordinates}

\item{y_limits}{Limits of the data in the y coordinates}
}
\description{
Fast Kernel Density Estimation in 2D
}
