% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_local_maxima}
\alias{find_local_maxima}
\title{Find Local Maxima}
\usage{
find_local_maxima(z, neigh_size)
}
\arguments{
\item{z}{Matrix of 2D density values}

\item{neigh_size}{Number of neighboring points to consider}
}
\description{
Find Local Maxima
}
