% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_unmix.R
\name{optimize.unmix}
\alias{optimize.unmix}
\title{Optimize Unmix
Unmix using the AUtoSpectral method to adapt fluorophore spectral signatures
on a per-cell basis.}
\usage{
optimize.unmix(
  raw.data,
  unmixed,
  spectra,
  pos.thresholds,
  optimize.fluors,
  variants,
  weights = numeric(0),
  weighted = FALSE,
  nthreads = 0,
  speed = "fast"
)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra. Matrix.}

\item{unmixed}{Initial unmixing to be optimized. Expectation is that the data
will derive from AutoSpectral and will have the autofluorescence component
removed. Format: cells (rows) x fluorophores (columns). Matrix.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{pos.thresholds}{Named numeric. The thresholds to be considered "positive"
for a fluorophore in the unnmixing. Expectation: create this by calling
\code{get.spectral.variants}, which generates the thresholds based on the 99.5th
percentile of the unstained sample.}

\item{optimize.fluors}{Vector of fluorophores to be optimized. Must match the
rownames in \code{spectra} and the colnames of \code{unmixed}.}

\item{variants}{List of spectral signature matrices for the variants of each
fluorophore. Format: Named list with elements fluorophores, each element being
a matrix of spectra in format variants (rows) by detectors (columns). Prepare
using \code{get.spectral.variants}.}

\item{weights}{Optional numeric vector of weights (one per fluorescent
detector). Default is \code{numeric(0)}, in which case OLS unmixing will be used.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm. Default is \code{FALSE} and will use OLS.}

\item{nthreads}{Numeric, number of threads to use for parallel processing.
Default is \code{0}, which will allow OpenMP to use all available cores.}

\item{speed}{Selector for the precision-speed trade-off. Options are the
default \code{fast}, which selects the best spectral fit per cell by updating the
predicted values for each fluorophore independently without repeating the
unnmixing, \code{medium} which uses a Woodbury-Sherman-Morrison rank-one updating
of the unnmixing matrix for better results and a moderate slow-down, or \code{slow},
which explicitly recomputes the unmixing matrix for each variant for maximum
precision.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Optimize Unmix
Unmix using the AUtoSpectral method to adapt fluorophore spectral signatures
on a per-cell basis.
}
