% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_optimization_inputs.R
\name{sanitize.optimization.inputs}
\alias{sanitize.optimization.inputs}
\title{Sanitize Optimization Inputs for C++}
\usage{
sanitize.optimization.inputs(spectra, optimize.fluors, variants, delta.norms)
}
\arguments{
\item{spectra}{Numeric matrix (fluors x detectors)}

\item{optimize.fluors}{Character vector of fluorophores present in variants}

\item{variants}{List of variant matrices per fluorophore}

\item{delta.norms}{List of delta norms per fluorophore}
}
\value{
Updated list of fluorophores to be optimized
}
\description{
Checks the delta list and delta norms for empty, NULL or redundant content,
removes fluorophores from the list of ones to be optimized if they fail checks.
The aim is prevent hard crashes in the C++ code, which can be triggered by NA,
NaN (divide by zero) or empty list elements.
}
