% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_irls.R
\name{poisson_irls_rcpp_parallel}
\alias{poisson_irls_rcpp_parallel}
\title{Poisson IRLS Parallel}
\usage{
poisson_irls_rcpp_parallel(
  raw_data,
  spectra,
  beta_init,
  maxit = 25L,
  tol = 1e-06,
  n_threads = 1L,
  divergence_threshold = 10000
)
}
\arguments{
\item{raw_data}{matrix of cells x detectors}

\item{spectra}{matrix of fluorophores x detectors}

\item{beta_init}{initial unmixed matrix of cells x fluorophores}

\item{maxit}{maximum number of IRLS iterations}

\item{tol}{convergence tolerance}

\item{n_threads}{number of threads to use (OpenMP)}

\item{divergence_threshold}{maximum allowed divergence before fallback}
}
\description{
Poisson IRLS Parallel
}
