% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_autospectral_rcpp.R
\name{unmix.autospectral.rcpp}
\alias{unmix.autospectral.rcpp}
\title{Unmix AutoSpectral Rcpp}
\usage{
unmix.autospectral.rcpp(
  raw.data,
  spectra,
  af.spectra,
  spectra.variants = NULL,
  weighted = FALSE,
  weights = NULL,
  calculate.error = FALSE,
  use.dist0 = TRUE,
  verbose = TRUE,
  parallel = TRUE,
  threads = 0,
  speed = "fast"
)
}
\arguments{
\item{raw.data}{Expression data from raw fcs files. Cells in rows and
detectors in columns. Columns should be fluorescent data only and must
match the columns in spectra.}

\item{spectra}{Spectral signatures of fluorophores, normalized between 0
and 1, with fluorophores in rows and detectors in columns.}

\item{af.spectra}{Spectral signatures of autofluorescences, normalized
between 0 and 1, with fluorophores in rows and detectors in columns. Prepare
using \code{get.af.spectra}.}

\item{spectra.variants}{Named list (names are fluorophores) carrying matrices
of spectral signature variations for each fluorophore. Prepare using
\code{get.spectral.variants}. Default is \code{NULL}.}

\item{weighted}{Logical, whether to use ordinary or weighted least squares
unmixing as the base algorithm. Default is \code{FALSE} and will use OLS.}

\item{weights}{Optional numeric vector of weights (one per fluorescent
detector). Default is \code{NULL}, in which case weighting will be done by
channel means (Poisson variance). Only used if \code{weighted}.}

\item{calculate.error}{Logical, whether to calculate the RMSE unmixing model
accuracy and include it as an output. Default is \code{FALSE}.}

\item{use.dist0}{Logical, controls whether the selection of the optimal AF
signature for each cell is determined by which unmixing brings the cell
closest to 0 (\code{use.dist0} = \code{TRUE}) or by which unmixing minimizes the
per-cell residual (\code{use.dist0} = \code{FALSE}). Default is \code{TRUE}.}

\item{verbose}{Logical, whether to send messages to the console.
Default is \code{TRUE}.}

\item{parallel}{Logical, whether to use parallel processing for the per-cell
unmixing for fluorophore optimization. Default is \code{FALSE}.}

\item{threads}{Numeric, number of threads to use for parallel processing.
Default is \code{0}, which will use all available cores. Recommended: use
\code{asp$worker.process.n}.}

\item{speed}{Selector for the precision-speed trade-off. Options are the
default \code{fast}, which selects the best spectral fit per cell by updating the
predicted values for each fluorophore independently without repeating the
unnmixing, \code{medium} which uses a Woodbury-Sherman-Morrison rank-one updating
of the unnmixing matrix for better results and a moderate slow-down, or \code{slow},
which explicitly recomputes the unmixing matrix for each variant for maximum
precision.}
}
\value{
Unmixed data with cells in rows and fluorophores in columns.
}
\description{
Unmix using the AutoSpectral method to extract autofluorescence and optimize
fluorophore signatures at the single cell level.
}
